# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Speech-AI Forge API",
    "description": "\n🍦 Speech-AI-Forge 是一个围绕 TTS 生成模型 ChatTTS 开发的项目，实现了 API Server 和 基于 Gradio 的 WebUI。<br/>\n🍦 Speech-AI-Forge is a project developed around the TTS generation model ChatTTS, implementing an API Server and a Gradio-based WebUI.\n\n项目地址: [https://github.com/lenML/Speech-AI-Forge](https://github.com/lenML/Speech-AI-Forge)\n\n> 所有生成音频的 POST api都无法在此页面调试，调试建议使用 playground <br/>\n> All audio generation POST APIs cannot be debugged on this page, it is recommended to use playground for debugging\n\n> 如果你不熟悉本系统，建议从这个一键脚本开始，在colab中尝试一下：<br/>\n> [https://colab.research.google.com/github/lenML/Speech-AI-Forge/blob/main/colab.ipynb](https://colab.research.google.com/github/lenML/Speech-AI-Forge/blob/main/colab.ipynb)\n\n[[Learn More About Documents]](https://github.com/lenML/Speech-AI-Forge/issues/240)\n            ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/tts": {
      "get": {
        "tags": [
          "TTS"
        ],
        "summary": "Synthesize Tts",
        "description": "**Text-to-Speech Synthesis API (v1 - GET)**\n\nThis endpoint converts text into speech using GET request parameters.\nIt offers various options for customizing the voice, style, audio output, and processing.\n\n**Mandatory Parameter**:\n*   `text`: The text string to be synthesized.\n\n**Speaker and Style Customization**:\n*   `spk`: Specify the speaker by name or seed (e.g., \"female2\").\n*   `style`: Define the speaking style (e.g., \"chat\").\n\n**Generation Control**:\n*   `temperature`: Sampling temperature (0.0-1.0, e.g., 0.3). Controls randomness.\n*   `top_p`: Nucleus sampling probability (0.0-1.0, e.g., 0.5).\n*   `top_k`: Limits sampling to the K most likely next tokens (e.g., 20).\n*   `seed`: Seed for reproducible generation (e.g., 42).\n*   `prompt`, `prompt1`, `prompt2`: Optional text prompts to guide inference.\n*   `prefix`: Optional text prefix for inference.\n*   `bs` (batch_size): Batch size for processing (e.g., \"8\").\n*   `thr` (threshold): Sentence splitter threshold (e.g., \"100\").\n*   `eos`: End-of-sentence marker (e.g., \"[uv_break]\").\n\n**Output Audio Configuration**:\n*   `format`: Desired audio output format. Supported: \"mp3\", \"wav\", \"raw\" (default).\n*   `bitrate`: Audio bitrate for compressed formats (e.g., \"64k\").\n\n**Audio Adjustments**:\n*   `speed`: Playback speed multiplier (e.g., 1.0 for normal).\n*   `pitch`: Pitch adjustment (e.g., 0 for no change).\n*   `volume_gain`: Volume gain in dB (e.g., 0 for no change).\n\n**Enhancements**:\n*   `enhance`: Boolean to enable audio enhancement (default: false).\n*   `denoise`: Boolean to enable audio denoising (default: false).\n\n**Streaming Output**:\n*   `stream`: Boolean to enable streaming audio generation (default: false).\n*   `chunk_size`: Size of chunks for streaming (e.g., 64, if stream is true).\n\n**Caching**:\n*   `no_cache`: Boolean or \"on\"/\"off\" to disable caching (default: false).\n\n**Model Selection**:\n*   `model`: Specify the TTS model ID to use (e.g., \"chat-tts\", \"cosy-voice\").\n\n**Response**:\n*   **Success**: An audio file stream (`FileResponse`).\n*   **Failure**: A JSON object detailing the error (e.g., validation errors, internal server error).\n\n**Note**: This v1 endpoint does *not* support voice cloning via reference audio. For voice cloning, please refer to the v2 API. Parameters like temperature, top_p, top_k, seed, prompts, and prefix might be overridden by specific speaker (`spk`) or style (`style`) configurations.",
        "operationId": "synthesize_tts_v1_tts_get",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text"
            }
          },
          {
            "name": "spk",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "female2",
              "title": "Spk"
            }
          },
          {
            "name": "style",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "chat",
              "title": "Style"
            }
          },
          {
            "name": "temperature",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.3,
              "title": "Temperature"
            }
          },
          {
            "name": "top_p",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.5,
              "title": "Top P"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Top K"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 42,
              "title": "Seed"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "raw",
              "title": "Format"
            }
          },
          {
            "name": "bitrate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "64k",
              "title": "Bitrate"
            }
          },
          {
            "name": "prompt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Prompt"
            }
          },
          {
            "name": "prompt1",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Prompt1"
            }
          },
          {
            "name": "prompt2",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Prompt2"
            }
          },
          {
            "name": "prefix",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Prefix"
            }
          },
          {
            "name": "bs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "8",
              "title": "Bs"
            }
          },
          {
            "name": "thr",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "100",
              "title": "Thr"
            }
          },
          {
            "name": "eos",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "[uv_break]",
              "title": "Eos"
            }
          },
          {
            "name": "enhance",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Enhance"
            }
          },
          {
            "name": "denoise",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Denoise"
            }
          },
          {
            "name": "speed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1.0,
              "title": "Speed"
            }
          },
          {
            "name": "pitch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0,
              "title": "Pitch"
            }
          },
          {
            "name": "volume_gain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0,
              "title": "Volume Gain"
            }
          },
          {
            "name": "stream",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Stream"
            }
          },
          {
            "name": "chunk_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 64,
              "title": "Chunk Size"
            }
          },
          {
            "name": "no_cache",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "enum": [
                    "on",
                    "off"
                  ],
                  "type": "string"
                }
              ],
              "default": false,
              "title": "No Cache"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "chat-tts",
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/xtts_v2/tts_to_audio": {
      "post": {
        "tags": [
          "XTTS"
        ],
        "summary": "Tts To Audio",
        "description": "[[Click To XTTS_V2 API Document]](https://github.com/lenML/Speech-AI-Forge/blob/main/docs/api_xtts.md)\n\n[[Learn More About Documents]](https://github.com/lenML/Speech-AI-Forge/issues/240)",
        "operationId": "tts_to_audio_v1_xtts_v2_tts_to_audio_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SynthesisParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/xtts_v2/tts_stream": {
      "get": {
        "tags": [
          "XTTS"
        ],
        "summary": "Tts Stream",
        "description": "[[Click To XTTS_V2 API Document]](https://github.com/lenML/Speech-AI-Forge/blob/main/docs/api_xtts.md)\n\n[[Learn More About Documents]](https://github.com/lenML/Speech-AI-Forge/issues/240)",
        "operationId": "tts_stream_v1_xtts_v2_tts_stream_get",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text"
            }
          },
          {
            "name": "speaker_wav",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Speaker Wav"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "cn",
              "title": "Language"
            }
          },
          {
            "name": "no_cache",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "No Cache"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/xtts_v2/set_tts_settings": {
      "post": {
        "tags": [
          "XTTS"
        ],
        "summary": "Set Tts Settings",
        "description": "[[Click To XTTS_V2 API Document]](https://github.com/lenML/Speech-AI-Forge/blob/main/docs/api_xtts.md)\n\n[[Learn More About Documents]](https://github.com/lenML/Speech-AI-Forge/issues/240)",
        "operationId": "set_tts_settings_v1_xtts_v2_set_tts_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/tts": {
      "post": {
        "tags": [
          "Forge V2"
        ],
        "summary": "Forge Text Synthesize",
        "description": "**Text-to-Speech Synthesis API (v2)**\n\nThis endpoint converts text into speech, offering a wide range of customization options.\nIt accepts JSON formatted data in the request body.\n\n**Core Functionality**:\n*   Synthesizes speech from various text inputs.\n*   Supports voice cloning by directly uploading a reference audio.\n\n**Input Flexibility**:\n*   Provide text via one of these mutually exclusive fields:\n    *   `text`: A single string of text.\n    *   `texts`: A list of strings for batch processing.\n    *   `ssml`: Text formatted using Speech Synthesis Markup Language (SSML).\n\n**Speaker Customization (`spk`)**:\n*   Use pre-defined speakers via `spk.from_spk_id` or `spk.from_spk_name`.\n*   **Voice Cloning**: Provide reference audio for voice cloning:\n    *   `spk.from_ref.wav_b64`: Base64 encoded WAV audio data.\n    *   `spk.from_ref.text`: The transcript corresponding to the reference audio.\n\n**Audio Processing & Control**:\n*   **Adjustments (`adjust`)**:\n    *   `pitch`: Modify audio pitch.\n    *   `speed_rate`: Adjust speaking rate.\n    *   `volume_gain_db`: Change audio volume.\n    *   `normalize`: Apply volume normalization.\n    *   `remove_silence`: Trim silence from audio ends.\n*   **Encoding (`encoder`)**:\n    *   `format`: Output audio format (e.g., \"mp3\", \"wav\").\n    *   `bitrate`: Audio bitrate (e.g., \"64k\").\n    *   `acodec`: Audio codec (e.g., \"libmp3lame\").\n*   **Enhancement (`enhance`)**:\n    *   `enabled`: Toggle audio enhancement.\n    *   `model`: Select enhancement model.\n    *   Additional parameters like `nfe`, `solver`, `lambd`, `tau` for fine-tuning.\n*   **Inference (`infer`)**:\n    *   `batch_size`, `spliter_threshold`, `eos` for text processing.\n    *   `seed`: For reproducible outputs.\n    *   `stream`, `stream_chunk_size`: Enable and configure streaming output.\n    *   `no_cache`, `sync_gen`: Control caching and generation mode.\n*   **Text Normalization (`tn`)**:\n    *   `enabled`/`disabled`: Specify text normalization rules.\n*   **TTS Model Parameters (`tts`)**:\n    *   `mid`: TTS model ID (e.g., \"cosy-voice\").\n    *   `style`, `emotion`: Control speaking style and emotion.\n    *   `temperature`, `top_p`, `top_k`: Adjust sampling parameters for generation.\n    *   Deprecated: `prompt`, `prompt1`, `prompt2`, `prefix`.\n*   **Voice Cloning (`vc`)** (Note: Potentially deprecated, mainly for OpenVoice):\n    *   `enabled`, `mid`, `emotion`, `tau`.\n\n**Response**:\n*   **Success**: An audio file stream (`FileResponse`).\n*   **Failure**: A JSON object detailing the error.\n\n**Key Considerations**:\n1.  Ensure only one of `text`, `texts`, or `ssml` is provided.\n2.  For voice cloning, use high-quality reference audio (3-5 seconds recommended).\n3.  Enabling audio enhancement may increase generation latency.",
        "operationId": "forge_text_synthesize_v2_tts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V2TtsParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SynthesisParams": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "speaker_wav": {
            "type": "string",
            "title": "Speaker Wav"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "cn"
          }
        },
        "type": "object",
        "required": [
          "text",
          "speaker_wav"
        ],
        "title": "SynthesisParams"
      },
...
      "TTSSettingsRequest": {
        "properties": {
          "stream_chunk_size": {
            "type": "integer",
            "title": "Stream Chunk Size"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature"
          },
          "speed": {
            "type": "number",
            "title": "Speed"
          },
          "length_penalty": {
            "type": "number",
            "title": "Length Penalty"
          },
          "repetition_penalty": {
            "type": "number",
            "title": "Repetition Penalty"
          },
          "top_p": {
            "type": "number",
            "title": "Top P"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K"
          },
          "enable_text_splitting": {
            "type": "boolean",
            "title": "Enable Text Splitting"
          },
          "batch_size": {
            "type": "integer",
            "title": "Batch Size"
          },
          "eos": {
            "type": "string",
            "title": "Eos"
          },
          "infer_seed": {
            "type": "integer",
            "title": "Infer Seed"
          },
          "use_decoder": {
            "type": "boolean",
            "title": "Use Decoder"
          },
          "prompt1": {
            "type": "string",
            "title": "Prompt1"
          },
          "prompt2": {
            "type": "string",
            "title": "Prompt2"
          },
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "spliter_threshold": {
            "type": "integer",
            "title": "Spliter Threshold"
          },
          "style": {
            "type": "string",
            "title": "Style"
          }
        },
        "type": "object",
        "required": [
          "stream_chunk_size",
          "temperature",
          "speed",
          "length_penalty",
          "repetition_penalty",
          "top_p",
          "top_k",
          "enable_text_splitting"
        ],
        "title": "TTSSettingsRequest"
      },
...
      "V2TtsParams": {
        "properties": {
          "adjust": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdjustConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "encoder": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EncoderConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "enhance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EnhancerConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "infer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "vc": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VCConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "tn": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TNConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "tts": {
            "$ref": "#/components/schemas/TTSConfig"
          },
          "spk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpeakerConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "texts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Texts"
          },
          "ssml": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssml"
          }
        },
        "type": "object",
        "title": "V2TtsParams"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:7870"
    },
    {
      "url": "http://0.0.0.0:7870"
    },
    {
      "url": "http://localhost:7870"
    }
  ]
}